refactor: move more xblock-data logic into xblock app
