+++ openedx/core/djangoapps/content_libraries/api/containers.py
@@ -26,8 +26,7 @@ from openedx_events.content_authoring.signals import (
 from openedx_learning.api import authoring as authoring_api
 from openedx_learning.api.authoring_models import Container, ContainerVersion, Component
 from openedx.core.djangoapps.content_libraries.api.collections import library_collection_locator
-
-from openedx.core.djangoapps.xblock.api import get_component_from_usage_key
+from openedx.core.djangoapps.xblock.api import create_xblock_field_data_for_container, get_component_from_usage_key
 
 from ..models import ContentLibrary
 from .exceptions import ContentLibraryContainerNotFound
@@ -413,146 +412,6 @@ def update_container(
         )
 
     return ContainerMetadata.from_container(library_key, version.container)
-
-
-from openedx.core.djangoapps.xblock.models import XBlockVersionFieldData
-
-def create_xblock_field_data_for_container(version: ContainerVersion):
-    # this whole thing should be in xblock.api instead of here.
-
-    log.info("Am I even being called?")
-
-    from openedx.core.djangoapps.xblock.models import Block
-
-    entity = version.publishable_entity_version.entity
-
-    # If this PublishableEntity isn't associated with an Learning Core backed
-    # XBlock, then we can't write anything. Note: This is going to be an edge
-    # case later, when we want to add an existing container to a container that
-    # was imported from a course.
-    if not hasattr(entity, 'block'):
-        log.error("No Block detected???")
-        return
-
-    parent_block = entity.block
-    container_usage_key = parent_block.key
-    course_key = container_usage_key.course_key
-
-    # Generic values for all container types
-    content_scoped_fields = {}
-    settings_scoped_fields = {
-        'display_name': version.publishable_entity_version.title
-    }
-    children = []
-
-    # Things specific to the course root...
-    if container_usage_key.block_type == "course":
-        content_scoped_fields['license'] = None
-        content_scoped_fields['wiki_slug'] = f'{course_key.org}.{course_key.course}.{course_key.run}'
-        settings_scoped_fields.update(
-            _course_block_entry(container_usage_key)
-        )
-
-    for child_entity_row in version.entity_list.entitylistrow_set.select_related('entity__block').all():
-        log.error(f"Iterating children: {child_entity_row.entity}")
-        if not hasattr(child_entity_row.entity, 'block'):
-            # This can happen if we add a new component in a library to a
-            # container that was imported from a course.
-            match(container_usage_key.block_type):
-                case "course":
-                    child_block_type = "chapter"
-                    child_block_id = child_entity_row.entity.key
-                case "chapter":
-                    child_block_type = "sequential"
-                    child_block_id = child_entity_row.entity.key
-                case "sequential":
-                    child_block_type = "vertical"
-                    child_block_id = child_entity_row.entity.key
-                case "vertical":
-                    child_block_type = child_entity_row.entity.component.component_type.name
-                    child_block_id = child_entity_row.entity.component.local_key
-
-            log.info(f"Creating child usage key: {child_usage_key}")
-            child_usage_key = course_key.make_usage_key(child_block_type, child_block_id)
-            child_block = Block.objects.create(
-                learning_context_id=parent_block.learning_context_id,
-                entity=child_entity_row.entity,
-                key=child_usage_key,
-            )
-        else:
-            child_block = child_entity_row.entity.block
-            child_usage_key = child_block.key
-        children.append(
-            [child_usage_key.block_type, child_usage_key.block_id]
-        )
-
-    field_data = XBlockVersionFieldData.objects.create(
-        pk=version.pk,
-        content=content_scoped_fields,
-        settings=settings_scoped_fields,
-        children=children,
-    )
-    log.info(f"Wrote XBlock Data for Container: {version}: {field_data}")
-
-
-def _course_block_entry(usage_key):
-    return {
-        'allow_anonymous': True,
-        'allow_anonymous_to_peers': False,
-        'cert_html_view_enabled': True,
-        'discussion_blackouts': [],
-        'discussion_topics': {'General': {'id': 'course'}},
-        'discussions_settings': {
-            'enable_graded_units': False,
-            'enable_in_context': True,
-            'openedx': { 'group_at_subsection': False},
-            'posting_restrictions': 'disabled',
-            'provider_type': 'openedx',
-            'unit_level_visibility': True
-        },
-        'end': None,
-        'language': 'en',
-
-        ## HARDCODED START DATE
-        'start': datetime(2020, 1, 1, 0, 0, tzinfo=timezone.utc),
-        'static_asset_path': 'course',
-        'tabs': [
-            {
-                'course_staff_only': False,
-                'name': 'Course',
-                'type': 'courseware'
-            },
-            {
-                'course_staff_only': False,
-                'name': 'Progress',
-                'type': 'progress'
-            },
-            {
-                'course_staff_only': False,
-                'name': 'Dates',
-                'type': 'dates'
-            },
-            {
-                'course_staff_only': False,
-                'name': 'Discussion',
-                'type': 'discussion'
-            },
-            {
-                'course_staff_only': False,
-                'is_hidden': True,
-                'name': 'Wiki',
-                'type': 'wiki'
-            },
-            {
-                'course_staff_only': False,
-                'name': 'Textbooks',
-                'type': 'textbooks'
-            }
-        ],
-        'xml_attributes': {
-            'filename': [ f'course/{usage_key.run}.xml', f'course/{usage_key.run}.xml']
-        }
-    }
 
 
 def delete_container(
