+++ openedx/core/djangoapps/xblock/api.py
@@ -18,7 +18,7 @@ from django.core.exceptions import PermissionDenied
 from django.urls import reverse
 from django.utils.translation import gettext as _
 from openedx_learning.api import authoring as authoring_api
-from openedx_learning.api.authoring_models import Component, ComponentVersion, PublishLog
+from openedx_learning.api.authoring_models import Component, ComponentVersion, ContainerVersion, PublishLog
 from opaque_keys.edx.keys import UsageKeyV2
 from opaque_keys.edx.locator import LibraryUsageLocatorV2
 from rest_framework.exceptions import NotFound
@@ -439,6 +439,144 @@ def handle_library_publish(publish_log: PublishLog):
     for course_key in affected_course_keys:
         log.info(f"Type of course_key: {type(course_key)}")
         update_learning_core_course(course_key)
+
+
+def create_xblock_field_data_for_container(version: ContainerVersion):
+    # this whole thing should be in xblock.api instead of here.
+
+    log.info("Am I even being called?")
+
+    from openedx.core.djangoapps.xblock.models import Block
+
+    entity = version.publishable_entity_version.entity
+
+    # If this PublishableEntity isn't associated with an Learning Core backed
+    # XBlock, then we can't write anything. Note: This is going to be an edge
+    # case later, when we want to add an existing container to a container that
+    # was imported from a course.
+    if not hasattr(entity, 'block'):
+        log.error("No Block detected???")
+        return
+
+    parent_block = entity.block
+    container_usage_key = parent_block.key
+    course_key = container_usage_key.course_key
+
+    # Generic values for all container types
+    content_scoped_fields = {}
+    settings_scoped_fields = {
+        'display_name': version.publishable_entity_version.title
+    }
+    children = []
+
+    # Things specific to the course root...
+    if container_usage_key.block_type == "course":
+        content_scoped_fields['license'] = None
+        content_scoped_fields['wiki_slug'] = f'{course_key.org}.{course_key.course}.{course_key.run}'
+        settings_scoped_fields.update(
+            _course_block_entry(container_usage_key)
+        )
+
+    for child_entity_row in version.entity_list.entitylistrow_set.select_related('entity__block').all():
+        log.error(f"Iterating children: {child_entity_row.entity}")
+        if not hasattr(child_entity_row.entity, 'block'):
+            # This can happen if we add a new component in a library to a
+            # container that was imported from a course.
+            match(container_usage_key.block_type):
+                case "course":
+                    child_block_type = "chapter"
+                    child_block_id = child_entity_row.entity.key
+                case "chapter":
+                    child_block_type = "sequential"
+                    child_block_id = child_entity_row.entity.key
+                case "sequential":
+                    child_block_type = "vertical"
+                    child_block_id = child_entity_row.entity.key
+                case "vertical":
+                    child_block_type = child_entity_row.entity.component.component_type.name
+                    child_block_id = child_entity_row.entity.component.local_key
+
+            log.info(f"Creating child usage key: {child_usage_key}")
+            child_usage_key = course_key.make_usage_key(child_block_type, child_block_id)
+            child_block = Block.objects.create(
+                learning_context_id=parent_block.learning_context_id,
+                entity=child_entity_row.entity,
+                key=child_usage_key,
+            )
+        else:
+            child_block = child_entity_row.entity.block
+            child_usage_key = child_block.key
+        children.append(
+            [child_usage_key.block_type, child_usage_key.block_id]
+        )
+
+    field_data = XBlockVersionFieldData.objects.create(
+        pk=version.pk,
+        content=content_scoped_fields,
+        settings=settings_scoped_fields,
+        children=children,
+    )
+    log.info(f"Wrote XBlock Data for Container: {version}: {field_data}")
+
+
+def _course_block_entry(usage_key):
+    return {
+        'allow_anonymous': True,
+        'allow_anonymous_to_peers': False,
+        'cert_html_view_enabled': True,
+        'discussion_blackouts': [],
+        'discussion_topics': {'General': {'id': 'course'}},
+        'discussions_settings': {
+            'enable_graded_units': False,
+            'enable_in_context': True,
+            'openedx': { 'group_at_subsection': False},
+            'posting_restrictions': 'disabled',
+            'provider_type': 'openedx',
+            'unit_level_visibility': True
+        },
+        'end': None,
+        'language': 'en',
+
+        ## HARDCODED START DATE
+        'start': datetime(2020, 1, 1, 0, 0, tzinfo=timezone.utc),
+        'static_asset_path': 'course',
+        'tabs': [
+            {
+                'course_staff_only': False,
+                'name': 'Course',
+                'type': 'courseware'
+            },
+            {
+                'course_staff_only': False,
+                'name': 'Progress',
+                'type': 'progress'
+            },
+            {
+                'course_staff_only': False,
+                'name': 'Dates',
+                'type': 'dates'
+            },
+            {
+                'course_staff_only': False,
+                'name': 'Discussion',
+                'type': 'discussion'
+            },
+            {
+                'course_staff_only': False,
+                'is_hidden': True,
+                'name': 'Wiki',
+                'type': 'wiki'
+            },
+            {
+                'course_staff_only': False,
+                'name': 'Textbooks',
+                'type': 'textbooks'
+            }
+        ],
+        'xml_attributes': {
+            'filename': [ f'course/{usage_key.run}.xml', f'course/{usage_key.run}.xml']
+        }
+    }
 
 
 class LearningCoreCourseShimWriter:
